Replace VisionSubsystem with PhotonCameraWrapper class
