+++ src/main/java/org/first5924/frc2023/constants/VisionConstants.java
@@ -9,6 +9,7 @@ import edu.wpi.first.math.geometry.Translation3d;
 
 /** Add your docs here. */
 public class VisionConstants {
+    public static final String kCameraName = "camera";
     public static final Translation3d kRobotToCamTranslation = new Translation3d(0, 0, 0);
     public static final Rotation3d kRobotToCamRotation = new Rotation3d(0, 0, 0);
 }
