+++ src/main/java/org/first5924/frc2023/robot/RobotContainer.java
@@ -7,10 +7,8 @@ package org.first5924.frc2023.robot;
 import org.first5924.frc2023.commands.autonomous.DriveOneMeter;
 import org.first5924.frc2023.commands.drive.CurvatureDrive;
 import org.first5924.frc2023.commands.drive.TurnInPlace;
-import org.first5924.frc2023.commands.vision.UpdatePoseEstimator;
 import org.first5924.frc2023.constants.OIConstants;
 import org.first5924.frc2023.subsystems.DriveSubsystem;
-import org.first5924.frc2023.subsystems.VisionSubsystem;
 
 import edu.wpi.first.wpilibj2.command.Command;
 import edu.wpi.first.wpilibj2.command.button.CommandXboxController;
@@ -25,7 +23,6 @@ import edu.wpi.first.wpilibj2.command.button.Trigger;
 public class RobotContainer {
   // The robot's subsystems and commands are defined here...
   private final DriveSubsystem mDrive = new DriveSubsystem();
-  private final VisionSubsystem mVision = new VisionSubsystem();
 
   // Replace with CommandPS4Controller or CommandJoystick if needed
   private final CommandXboxController mDriverController =
@@ -34,7 +31,6 @@ public class RobotContainer {
   /** The container for the robot. Contains subsystems, OI devices, and commands. */
   public RobotContainer() {
     mDrive.setDefaultCommand(new CurvatureDrive(mDrive, mDriverController::getLeftY, mDriverController::getRightX));
-    mVision.setDefaultCommand(new UpdatePoseEstimator(mDrive, mVision));
 
     // Configure the trigger bindings
     configureBindings();
